Decouple the data from the output
Decouple the data by introducing data structs that contains the parsed
data that can be passed to an output implementation.
This only includes a plain text output implementation, with a not yet
stable format.
This work will enable different output types more easily.
Also the data structs can be used for the upcoming diff command.